PR feedback, consolidate to list_blobs own file
