+++ sdk/storage/azure_storage_blob/assets.json
@@ -1,6 +1,6 @@
 {
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "rust",
-  "Tag": "rust/azure_storage_blob_55bea387aa",
+  "Tag": "rust/azure_storage_blob_c202ef32f3",
   "TagPrefix": "rust/azure_storage_blob"
 }
